fix: team subscription
